+++ src/components/Profiling/Dialog/CompleteDialog.tsx
@@ -124,7 +124,7 @@ const SaveBatteryModelButton = ({
 
     const finishedProfiles =
         project?.settings?.profiles.filter(
-            prof => prof.batteryJson && prof.paramsJson
+            prof => prof.batteryJson && prof.paramsJson && prof.batteryInc
         ) ?? [];
 
     return (
@@ -142,6 +142,10 @@ const SaveBatteryModelButton = ({
                             name: 'JSON',
                             extensions: ['json'],
                         },
+                        {
+                            name: 'INC',
+                            extensions: ['inc'],
+                        },
                     ],
                 }).then(result => {
                     if (
@@ -152,8 +156,11 @@ const SaveBatteryModelButton = ({
                         onGeneratingBatteryModel(true);
                         mergeBatteryParams(project.settings, finishedProfiles)
                             .then(data => {
-                                if (result.filePath)
-                                    stringToFile(result.filePath, data);
+                                if (result.filePath?.endsWith('.json')) {
+                                    stringToFile(result.filePath, data.json);
+                                } else if (result.filePath?.endsWith('.inc')) {
+                                    stringToFile(result.filePath, data.inc);
+                                }
                             })
                             .finally(() => {
                                 dispatch(finishProfiling());